+++ app/src/main/java/org/hyperledger/besu/cli/BesuCommand.java
@@ -2617,8 +2617,8 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
       if (subStorageConfiguration.getLimitTrieLogsEnabled()) {
         builder
             .setLimitTrieLogsEnabled()
-            .setTrieLogRetentionLimit(subStorageConfiguration.getMaxLayersToLoad())
-            .setTrieLogsPruningWindowSize(subStorageConfiguration.getTrieLogPruningWindowSize());
+            .setTrieLogRetentionLimit(subStorageConfiguration.getTrieLogRetentionLimit())
+            .setTrieLogsPruningBatchSize(subStorageConfiguration.getTrieLogPruningBatchSize());
       }
     }